refactor: 멤버 레포지토리 패키지 변경
